disable GHC environment files in another way
Although we may not want to create GHC environment files in a standard
location after every build, it is still useful to be able to build them
in a custom place in special circumstances. So let's disable this
feature at the call-sites where it doesn't work properly rather than
inside the implementation of the feature.